+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/IIP-Design/styled-block-builder/compare/v2.0.1...HEAD)
+## [Unreleased](https://github.com/IIP-Design/styled-block-builder/compare/v2.0.2...HEAD)
 
 _This sections lists changes committed since most recent release_
+
+## [v2.0.2](https://github.com/IIP-Design/styled-block-builder/compare/v2.0.1...v2.0.2) - 2020.09.24
 
 ### Added: